+++ _build/build.config.php
@@ -1,2 +1,1 @@
 <?php
-require_once(dirname(dirname(__FILE__)).'/config.core.php');