+++ config.core.php
@@ -1,3 +1,3 @@
 <?php
-define('MODX_CORE_PATH', '/Applications/MAMP/htdocs/modx-stable2/core/');
+define('MODX_CORE_PATH', '/Users/wayneroddy/Sites/modx-themes/modx-install/core/');
 define('MODX_CONFIG_KEY', 'config');